refactor(reviewer-picker): extract subcomponents and fix design lints
Split ReviewerPickerModal (727 lines, over 600 budget) into RoleBadge
and SelectedRowCard. Migrate net-new gray-* tailwind classes to zinc
and collapse w-N h-N redundant axes to size-N to clear quality-gate
react-doctor regressions.
